Added band_utility_funs, minor changes to model and workflow
The model now uses a normal prior and a logit link to allow for
covariates in the future.

The band_utility_funs include a subset function, a function to build
encounter histories, a function to fix errandt sex entries and a
function to generate inits.

The workflow is complete and functions well over a variety of species,
but comprehensive testing has not been done.

The markdown document is now complete in terms of explorations and
initial data handling.